# edwards: avoid inversions in Add in extended points