# Description

This PR proposes solving a performance problem while doing `Add` in Edwards curves. 

The TL;DR of the problem is that this operation does an equality checking, which involves a very slow inversion.


## Type of change

- [x]  Performance improvement

# How has this been tested?

This PR doesn’t change the logic or add new border cases, so existing tests cover correct behavior.

# How has this been benchmarked?

```
name                    old time/op    new time/op    delta
MixedAdd/Extended-16       221ns ± 2%     220ns ± 2%      ~     (p=0.393 n=10+10)
Add/Extended-16           3.70µs ± 1%    0.20µs ± 2%   -94.52%  (p=0.000 n=10+10)

name                    old alloc/op   new alloc/op   delta
MixedAdd/Extended-16       0.00B          0.00B           ~     (all equal)
Add/Extended-16            0.00B          0.00B           ~     (all equal)

name                    old allocs/op  new allocs/op  delta
MixedAdd/Extended-16        0.00           0.00           ~     (all equal)
Add/Extended-16             0.00           0.00           ~     (all equal)
```